Take the ClickHouse column types from config instead of the server
The sink read `system.columns` before every insert to a table it had not
seen, and encoded against whatever came back. But envio creates these
tables — it already knows their shape — and paying a round trip to be told
it back bought a second source of truth that could disagree with the first,
plus a schema cache to invalidate, a retry path, and a parser sitting in
front of every insert.

`stage` now carries each column's ClickHouse type alongside its values: the
same string the `CREATE TABLE` used, produced by the same function. That
alone would trade one fragility for another, because RowBinary is positional
and a table whose columns moved would be silently misread. So the insert now
names every column it sends:

    INSERT INTO db.tbl (`id`, `n`) FORMAT RowBinary

which is what makes taking the types from config safe, and is better than
what reading the server gave us:

- The table's own column order stops mattering.
- A column envio does not write is the server's to fill, so a DEFAULT
  expression is evaluated instead of being replaced by the type's zero
  value, and MATERIALIZED/ALIAS/EPHEMERAL need no special casing.
- A column whose type this encoder cannot represent — a UUID or Map someone
  added — is no longer parsed at all, where before it failed every insert to
  that table for the life of the process.

Enum variants are now numbered explicitly in the DDL. RowBinary carries the
number rather than the name, and ClickHouse assigns numbers to an unnumbered
list itself; writing them down makes the type text say what the encoder
needs rather than leaving it to a server rule.

Gone with the lookup: the schema cache and its invalidation, the read-query
retry, the defaulted-column warning, and `align_to_schema`. The mock server
loses its `system.columns` half and now counts read queries so a test can
assert the sink makes none.

+++ packages/cli/src/clickhouse/row_binary.rs
@@ -568,7 +568,7 @@ mod tests {
         let spans = spans_from_utf16_lengths(&data, &[1, 1, 2, 1]).unwrap();
         let values: Vec<&str> = spans
             .iter()
-            .map(|(s, e)| &data[*s as usize..*e as usize])
+            .map(|(s, e)| &data[*s..*e])
             .collect();
         assert_eq!(values, vec!["a", "é", "😀", "b"]);
     }
